Many bug fixes
