+++ conf_example.toml
@@ -27,8 +27,8 @@ export = "/mnt/shared"
 
 # This is for dpfs_rvfs_dpu and the dpfs_hal implementation that uses RVFS
 [rvfs]
-remote_uri = "10.100.0.1:38510"
-dpu_uri = "10.100.0.115:38510"
+remote_uri = "10.100.0.1:31850"
+dpu_uri = "10.100.0.115:31850"
 # Useful for the DPU and remote side to have the QD synced
 # Must be a power of 2 and >= 1
 queue_depth = 64
