+++ dpfs_rvfs/rvfs.h
@@ -1,0 +1,8 @@
+#ifndef DPFS_RVFS_H
+#define DPFS_RVFS_H
+
+#define DPFS_RVFS_MAX_REQRESP_SIZE ((2 << 20) + 4096)
+
+#define DPFS_RVFS_REQTYPE_FUSE 0
+
+#endif // DPFS_RVFS_H
